animals refactor

+++ src/animals/Crocodile.java
@@ -2,7 +2,7 @@ package animals;
 
 public class Crocodile extends Wild {
 
-    public Crocodile(int id, int age, int weight, String color, boolean isPredator) {
+    public Crocodile(int id, int age, double weight, String color, boolean isPredator) {
         super(id, age, weight, color, isPredator);
     }
 }
