+++ src/animals/Giraffe.java
@@ -2,7 +2,7 @@ package animals;
 
 public class Giraffe extends Wild{
 
-    public Giraffe(int id, int age, int weight, String color, boolean isPredator) {
+    public Giraffe(int id, int age, double weight, String color, boolean isPredator) {
         super(id, age, weight, color, isPredator);
     }
 }
