+++ src/animals/Hamster.java
@@ -2,7 +2,7 @@ package animals;
 
 public class Hamster extends Pet {
 
-    public Hamster(int id, int age, int weight, String color, String name, boolean isVaccinated) {
+    public Hamster(int id, int age, double weight, String color, String name, boolean isVaccinated) {
         super(id, age, weight, color, name, isVaccinated);
     }
 }
