+++ src/animals/Lion.java
@@ -2,7 +2,7 @@ package animals;
 
 public class Lion extends Wild {
 
-    public Lion(int id, int age, int weight, String color, boolean isPredator) {
+    public Lion(int id, int age, double weight, String color, boolean isPredator) {
         super(id, age, weight, color, isPredator);
     }
 }
