+++ src/animals/Volf.java
@@ -2,7 +2,7 @@ package animals;
 
 public class Volf extends Wild {
 
-    public Volf(int id, int age, int weight, String color, boolean isPredator) {
+    public Volf(int id, int age, double weight, String color, boolean isPredator) {
         super(id, age, weight, color, isPredator);
     }
 }
